devide radioButton in SettingDialog
